requirements.txt
mutagen==1.47.0
click==8.1.7
rich==13.7.1
Pillow==10.3.0
pyinstaller==6.5.0
# pygame - if needed for audio playing in future
# PyObjC is macOS only, removed for Windows support
